refactor: relocate and restructure GraphQL types and mutations into a public directory.
