+++ codegen.json
@@ -189,9 +189,13 @@
           "PagePreview": "../../postgres/types/index#PagePreview as PagePreviewDB",
           "PersistGitHubSearchQuerySuccess": "./types/PersistGitHubSearchQuerySuccess#PersistGitHubSearchQuerySuccessSource",
           "PersistJiraSearchQuerySuccess": "./types/PersistJiraSearchQuerySuccess#PersistJiraSearchQuerySuccessSource",
+          "PokerAnnounceDeckHoverSuccess": "./types/PokerAnnounceDeckHoverSuccess#PokerAnnounceDeckHoverSuccessSource",
           "PokerMeeting": "../../postgres/types/Meeting#PokerMeeting as PokerMeetingDB",
           "PokerMeetingMember": "../../postgres/types/Meeting.d#PokerMeetingMember as PokerMeetingMemberDB",
           "PokerMeetingSettings": "../../postgres/types/index#PokerMeetingSettings as PokerMeetingSettingsDB",
+          "PokerResetDimensionSuccess": "./types/PokerResetDimensionSuccess#PokerResetDimensionSuccessSource",
+          "PokerRevealVotesSuccess": "./types/PokerRevealVotesSuccess#PokerRevealVotesSuccessSource",
+          "PokerTemplateDimensionUpdateDescriptionPayload": "./types/PokerTemplateDimensionUpdateDescriptionPayload#PokerTemplateDimensionUpdateDescriptionPayloadSource",
           "RRule": "rrule-rust#RRuleSet",
           "Reactable": "../../postgres/types#Reactable as ReactableDB",
           "Reactji": "./types/Reactji#ReactjiSource",
@@ -199,7 +203,13 @@
           "RetroDiscussStage": "./types/RetroDiscussStage#DiscussStageSource",
           "ReflectTemplatePromptUpdateDescriptionPayload": "./types/ReflectTemplatePromptUpdateDescriptionPayload#ReflectTemplatePromptUpdateDescriptionPayloadSource",
           "ReflectTemplatePromptUpdateGroupColorPayload": "./types/ReflectTemplatePromptUpdateGroupColorPayload#ReflectTemplatePromptUpdateGroupColorPayloadSource",
+          "PromoteNewMeetingFacilitatorPayload": "./types/PromoteNewMeetingFacilitatorPayload#PromoteNewMeetingFacilitatorPayloadSource",
+          "PromoteToTeamLeadPayload": "./types/PromoteToTeamLeadPayload#PromoteToTeamLeadPayloadSource",
+          "PushInvitationPayload": "./types/PushInvitationPayload#PushInvitationPayloadSource",
           "RemoveAgendaItemPayload": "./types/RemoveAgendaItemPayload#RemoveAgendaItemPayloadSource",
+          "RemoveAtlassianAuthPayload": "./types/RemoveAtlassianAuthPayload#RemoveAtlassianAuthPayloadSource",
+          "RemoveGitHubAuthPayload": "./types/RemoveGitHubAuthPayload#RemoveGitHubAuthPayloadSource",
+          "RemoveIntegrationProviderSuccess": "./types/RemoveIntegrationProviderSuccess#RemoveIntegrationProviderSuccessSource",
           "RemoveApprovedOrganizationDomainsSuccess": "./types/RemoveApprovedOrganizationDomainsSuccess#RemoveApprovedOrganizationDomainsSuccessSource",
           "RemoveIntegrationSearchQuerySuccess": "./types/RemoveIntegrationSearchQuerySuccess#RemoveIntegrationSearchQuerySuccessSource",
           "RemoveOrgUsersSuccess": "./types/RemoveOrgUsersSuccess#RemoveOrgUsersSuccessSource",
