+++ packages/server/graphql/public/mutations/removeIntegrationProvider.ts
@@ -1,0 +1,72 @@
+import IntegrationProviderId from 'parabol-client/shared/gqlIds/IntegrationProviderId'
+import {SubscriptionChannel} from 'parabol-client/types/constEnums'
+import getKysely from '../../../postgres/getKysely'
+import {getUserId, isSuperUser, isTeamMember, isUserOrgAdmin} from '../../../utils/authorization'
+import publish from '../../../utils/publish'
+import standardError from '../../../utils/standardError'
+import type {MutationResolvers} from '../resolverTypes'
+
+const removeIntegrationProvider: MutationResolvers['removeIntegrationProvider'] = async (
+  _source,
+  {providerId},
+  {authToken, dataLoader, socketId: mutatorId}
+) => {
+  const viewerId = getUserId(authToken)
+  const operationId = dataLoader.share()
+  const subOptions = {mutatorId, operationId}
+
+  // AUTH
+  const providerDbId = IntegrationProviderId.split(providerId)
+  const provider = await dataLoader.get('integrationProviders').load(providerDbId)
+  if (!provider) return standardError(new Error('Integration Provider not found'))
+  const {teamId, orgId, scope} = provider
+
+  if (!isSuperUser(authToken)) {
+    if (scope === 'global') {
+      return {error: {message: 'Must be a super user to remove a global provider'}}
+    }
+    if (scope === 'org' && !(await isUserOrgAdmin(viewerId, orgId!, dataLoader))) {
+      return {error: {message: 'Must be a member of the organization that created the provider'}}
+    }
+    if (scope === 'team' && !isTeamMember(authToken, teamId!)) {
+      const team = await dataLoader.get('teams').load(teamId!)
+      if (!team || !(await isUserOrgAdmin(viewerId, team.orgId, dataLoader))) {
+        return {error: {message: 'Must be on the team that created the provider'}}
+      }
+    }
+  }
+
+  // RESOLUTION
+  await getKysely()
+    .with('removedTokens', (eb) =>
+      eb
+        .updateTable('TeamMemberIntegrationAuth')
+        .set({isActive: false})
+        .where('providerId', '=', providerDbId)
+        .where('isActive', '=', true)
+    )
+    .updateTable('IntegrationProvider')
+    .set({isActive: false})
+    .where('id', '=', providerDbId)
+    .where('isActive', '=', true)
+    .execute()
+
+  const data = {
+    providerId,
+    orgIntegrationProviders: orgId ? {orgId} : null,
+    teamMemberIntegrations: teamId ? {teamId, userId: viewerId} : null
+  }
+
+  if (orgId) {
+    publish(
+      SubscriptionChannel.ORGANIZATION,
+      orgId,
+      'RemoveIntegrationProviderSuccess',
+      data,
+      subOptions
+    )
+  }
+  return data
+}
+
+export default removeIntegrationProvider
