--- packages/server/graphql/types/PokerTemplateDimensionUpdateDescriptionPayload.ts
@@ -1,22 +1,0 @@
-import {GraphQLObjectType} from 'graphql'
-import type {GQLContext} from '../graphql'
-import StandardMutationError from './StandardMutationError'
-import TemplateDimension from './TemplateDimension'
-
-const PokerTemplateDimensionUpdateDescriptionPayload = new GraphQLObjectType<any, GQLContext>({
-  name: 'PokerTemplateDimensionUpdateDescriptionPayload',
-  fields: () => ({
-    error: {
-      type: StandardMutationError
-    },
-    dimension: {
-      type: TemplateDimension,
-      resolve: ({dimensionId}, _args: unknown, {dataLoader}) => {
-        if (!dimensionId) return null
-        return dataLoader.get('templateDimensions').load(dimensionId)
-      }
-    }
-  })
-})
-
-export default PokerTemplateDimensionUpdateDescriptionPayload
